fix(cluster): dev cluster is up
I have the dev cluster working, cilium patch did not work with
terraform, had to manually apply it afterwards in the boostrap script.

I'm okay with this as it could be run in ci/cd in a pinch but it's
brittle. I do think it's a bug in the provider not Talos itself.